// packages/api/src/server/analytics/entity.ts
export class ActivityStream {
  propertyIdentifier: string;

  action: string;

  props: Props;

  message: string;

  payload: string;

  source: string;

  createdBy: string;
}

export class Props {
  env: string;

  applicationIdentifier: string;

  cluster: string;

  type: string;
}

export enum Action {
  APPLICATION_DEPLOYED = 'APPLICATION_DEPLOYED',
  APPLICATION_DEPLOYMENT_STARTED = 'APPLICATION_DEPLOYMENT_STARTED',
  APPLICATION_DEPLOYMENT_PROCESSING = 'APPLICATION_DEPLOYMENT_PROCESSING',
  APPLICATION_DEPLOYMENT_FAILED = 'APPLICATION_DEPLOYMENT_FAILED',
  APPLICATION_DEPLOYMENT_TIMEOUT = 'APPLICATION_DEPLOYMENT_TIMEOUT',
  APPLICATION_DELETED = 'APPLICATION_DELETED',
  APPLICATION_CONFIG_UPDATED = 'APPLICATION_CONFIG_UPDATED',
  APPLICATION_BUILD_STARTED = 'APPLICATION_BUILD_STARTED',
  APPLICATION_BUILD_FINISHED = 'APPLICATION_BUILD_FINISHED',
  APPLICATION_BUILD_FAILED = 'APPLICATION_BUILD_FAILED',
  APPLICATION_BUILD_TIMEOUT = 'APPLICATION_BUILD_TIMEOUT',
  APPLICATION_BUILD_TERMINATED = 'APPLICATION_BUILD_TERMINATED',
  APPLICATION_UNAVAILABLE = 'APPLICATION_UNAVAILABLE',
  APPLICATION_AVAILABLE = 'APPLICATION_AVAILABLE',
  APIKEY_CREATED = 'APIKEY_CREATED',
  APIKEY_DELETED = 'APIKEY_DELETED',
  PROPERTY_CREATED = 'PROPERTY_CREATED',
  PROPERTY_UPDATED = 'PROPERTY_UPDATED',
  PROPERTY_DELETED = 'PROPERTY_DELETED',
  ENV_CREATED = 'ENV_CREATED',
  ENV_UPDATED = 'ENV_UPDATED',
  ENV_DELETED = 'ENV_DELETED',
  ENV_SYNCED = 'ENV_SYNCED',
  SYMLINK_CREATED = 'SYMLINK_CREATED',
  WEBHOOK_CREATED = 'WEBHOOK_CREATED',
  WEBHOOK_DELETED = 'WEBHOOK_DELETED',
  WEBHOOK_UPDATED = 'WEBHOOK_UPDATED',
  PERMISSION_CREATED = 'PERMISSION_CREATED',
  PERMISSION_UPDATED = 'PERMISSION_UPDATED',
  PERMISSION_DELETED = 'PERMISSION_DELETED',
  LIGHTHOUSE_REPORT_GENERATION_STARTED = 'LIGHTHOUSE_REPORT_GENERATION_STARTED',
  LIGHTHOUSE_REPORT_GENERATION_FAILED = 'LIGHTHOUSE_REPORT_GENERATION_FAILED',
  LIGHTHOUSE_REPORT_AUTO_GENERATION_ENABLED = 'LIGHTHOUSE_REPORT_AUTO_GENERATION_ENABLED',
  LIGHTHOUSE_REPORT_AUTO_GENERATION_DISABLED = 'LIGHTHOUSE_REPORT_AUTO_GENERATION_DISABLED'
}